+++ source/workers/Worker.h
@@ -37,8 +37,8 @@ class Worker
 		ProgArgs* progArgs; // shortcut for member of workersSharedData
 		size_t workerRank; // rank of this worker in range 0 to numWorkers-1
 
-		std::atomic_bool phaseFinished; /* true after finishPhase() until resetStats() to prevent
-			finishPhase() inc'ing done counter twice on interrupt in waitForNextPhase() */
+		std::atomic_bool phaseFinished{false}; /* true after finishPhase() until resetStats() to
+			prevent finishPhase() inc'ing done counter twice on interrupt in waitForNextPhase() */
 		UInt64Vec elapsedUSecVec; /* Microsecs. Only valid when phase completed successfully. For
 			LocalWorker: finish of only thread; for RemoteWorker: finish of each worker on host */
 		std::atomic_bool isInterruptionRequested{false}; // set true to request self-termination
@@ -47,7 +47,7 @@ class Worker
 		AtomicLiveOps oldAtomicLiveOps; // copy of old atomicLiveOps for diff stats
 		AtomicLiveOps oldAtomicLiveOpsReadMix; // copy of old atomicLiveOps for diff stats
 		std::atomic_bool stoneWallTriggered{false}; // true after 1st worker triggered stonewall
-		bool workerGotPhaseWork{true}; /* workers set this to false if they got no work
+		std::atomic_bool workerGotPhaseWork{true}; /* workers set this to false if they got no work
 			assigned and thus finish immediately. these also don't trigger stonewall. */
 		LiveOps stoneWallOps; // done values when stonewall was hit
 		LiveOps stoneWallOpsReadMix; // done values when stonewall was hit